Removed old file and renamed up to date version of calc_localgci.py
