+++ data.pickle
@@ -1,0 +1,309 @@
+(dp0
+S'Adams'
+p1
+cnumpy.core.multiarray
+_reconstruct
+p2
+(cnumpy
+ndarray
+p3
+(I0
+tp4
+S'b'
+p5
+tp6
+Rp7
+(I1
+(I32
+tp8
+cnumpy
+dtype
+p9
+(S'f8'
+p10
+I0
+I1
+tp11
+Rp12
+(I3
+S'<'
+p13
+NNNI-1
+I-1
+I0
+tp14
+bI00
+S'\x00\x00\x00\x00\x00\x00\xf0\xbf\x00\x00\x00\x00\x00\x00$\xc0\x00\x00\x00\x00\x00\x00&\xc0\x00\x00\x00\x00\x00\x00(\xc0\x00\x00\x00\x00\x00\x00*\xc0\x00\x00\x00\x00\x00\x00,\xc0\x00\x00\x00\x00\x00\x00.\xc0\x00\x00\x00\x00\x00\x000\xc0\x00\x00\x00\x00\x00\x001\xc0\x00\x00\x00\x00\x00\x002\xc0\x00\x00\x00\x00\x00\x003\xc0\x00\x00\x00\x00\x00\x00\x00\xc0\x00\x00\x00\x00\x00\x004\xc0\x00\x00\x00\x00\x00\x005\xc0\x00\x00\x00\x00\x00\x006\xc0\x00\x00\x00\x00\x00\x007\xc0\x00\x00\x00\x00\x00\x008\xc0\x00\x00\x00\x00\x00\x009\xc0\x00\x00\x00\x00\x00\x00:\xc0\x00\x00\x00\x00\x00\x00;\xc0\x00\x00\x00\x00\x00\x00<\xc0\x00\x00\x00\x00\x00\x00=\xc0\x00\x00\x00\x00\x00\x00\x08\xc0\x00\x00\x00\x00\x00\x00>\xc0\x00\x00\x00\x00\x00\x00?\xc0\x00\x00\x00\x00\x00\x00@\xc0\x00\x00\x00\x00\x00\x00\x10\xc0\x00\x00\x00\x00\x00\x00\x14\xc0\x00\x00\x00\x00\x00\x00\x18\xc0\x00\x00\x00\x00\x00\x00\x1c\xc0\x00\x00\x00\x00\x00\x00 \xc0\x00\x00\x00\x00\x00\x00"\xc0'
+p15
+tp16
+bsS'volumes'
+p17
+(lp18
+F11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+aF11.494040321933852
+asS'N'
+p19
+(dp20
+I0
+(lp21
+F0.7811550151975684
+aF0.9179331306990881
+aF0.547112462006079
+aF0.3434650455927052
+aF0.11246200607902736
+aF0.04559270516717325
+aF0.00303951367781155
+aF0.00303951367781155
+aF0.0
+aF0.0
+aF0.0
+aF0.9027355623100304
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.9544072948328267
+aF0.0
+aF0.0
+aF0.0
+aF0.9574468085106383
+aF0.9908814589665653
+aF0.9878419452887538
+aF0.9574468085106383
+aF0.9848024316109423
+aF0.9665653495440729
+asI1
+(lp22
+F0.7537993920972644
+aF0.9817629179331308
+aF0.756838905775076
+aF0.7021276595744681
+aF0.45896656534954405
+aF0.14893617021276595
+aF0.1094224924012158
+aF0.0182370820668693
+aF0.0121580547112462
+aF0.0
+aF0.0
+aF0.8054711246200608
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.9817629179331308
+aF0.0
+aF0.0
+aF0.0
+aF0.9969604863221885
+aF0.9240121580547113
+aF0.8875379939209727
+aF0.9057750759878419
+aF0.9604863221884499
+aF0.9878419452887538
+asI2
+(lp23
+F0.7933130699088146
+aF0.8753799392097265
+aF0.46504559270516715
+aF0.4133738601823708
+aF0.0911854103343465
+aF0.00911854103343465
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.5835866261398176
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.7507598784194529
+aF0.0
+aF0.0
+aF0.0
+aF0.6443768996960486
+aF0.8419452887537994
+aF0.8996960486322189
+aF0.9544072948328267
+aF0.9544072948328267
+aF0.9604863221884499
+asI3
+(lp24
+F0.5197568389057751
+aF0.47720364741641336
+aF0.24620060790273557
+aF0.0547112462006079
+aF0.0243161094224924
+aF0.0
+aF0.00303951367781155
+aF0.02127659574468085
+aF0.015197568389057751
+aF0.0121580547112462
+aF0.00303951367781155
+aF0.8632218844984803
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.0
+aF0.9604863221884499
+aF0.0
+aF0.0
+aF0.0
+aF0.9118541033434651
+aF0.9817629179331308
+aF0.9270516717325228
+aF0.9057750759878419
+aF0.939209726443769
+aF0.723404255319149
+ass.
